rename auth composables to match naming convention
